# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopClone E-Commerce API",
    "description": "A full-featured e-commerce REST API powering the ShopClone storefront. Provides endpoints for user authentication, product catalog browsing, shopping cart management, order processing, wishlists, product reviews, and simulated payment processing.",
    "version": "1.0.0",
    "contact": {
      "name": "ShopClone API Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List all products",
        "description": "Retrieves a list of all products in the catalog. Supports optional filtering by category and text search via query parameters.",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Filter products by category name",
            "schema": {
              "type": "string"
            },
            "example": "electronics"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Search query to filter products by title or description",
            "schema": {
              "type": "string"
            },
            "example": "wireless headphones"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                },
                "example": {
                  "products": [
                    {
                      "id": "prod-001",
                      "title": "Wireless Bluetooth Headphones",
                      "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                      "price": 79.99,
                      "image": "https://example.com/images/headphones.jpg",
                      "category": "electronics"
                    },
                    {
                      "id": "prod-002",
                      "title": "Organic Cotton T-Shirt",
                      "description": "Comfortable and sustainable organic cotton t-shirt available in multiple colors.",
                      "price": 24.99,
                      "image": "https://example.com/images/tshirt.jpg",
                      "category": "clothing"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Seed the product catalog",
        "description": "Seeds the product catalog with initial data. This endpoint is intended for development and setup purposes. The request body must include an action field set to \"seed\".",
        "operationId": "seedProducts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedProductsRequest"
              },
              "example": {
                "action": "seed"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products successfully seeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedProductsResponse"
                },
                "example": {
                  "success": true,
                  "message": "Successfully seeded 20 products"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing or invalid action field)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid action. Only 'seed' is supported."
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get a product by ID",
        "description": "Retrieves the details of a single product by its unique identifier.",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the product",
            "schema": {
              "type": "string"
            },
            "example": "prod-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleProductResponse"
                },
                "example": {
                  "product": {
                    "id": "prod-001",
                    "title": "Wireless Bluetooth Headphones",
                    "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                    "price": 79.99,
                    "image": "https://example.com/images/headphones.jpg",
                    "category": "electronics"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid product ID"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Product not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List all orders for the current user",
        "description": "Retrieves a list of all orders placed by the authenticated user, sorted by date in descending order.",
        "operationId": "listOrders",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved order list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                },
                "example": {
                  "orders": [
                    {
                      "id": "ord-abc123",
                      "items": [
                        {
                          "product": {
                            "id": "prod-001",
                            "title": "Wireless Bluetooth Headphones",
                            "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                            "price": 79.99,
                            "image": "https://example.com/images/headphones.jpg",
                            "category": "electronics"
                          },
                          "quantity": 2
                        }
                      ],
                      "total": 159.98,
                      "date": "2026-02-23T14:30:00.000Z",
                      "status": "pending"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Place a new order",
        "description": "Creates a new order for the authenticated user with the specified items and total amount. The order is initially created with a \"pending\" status.",
        "operationId": "createOrder",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              },
              "example": {
                "items": [
                  {
                    "product": {
                      "id": "prod-001",
                      "title": "Wireless Bluetooth Headphones",
                      "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                      "price": 79.99,
                      "image": "https://example.com/images/headphones.jpg",
                      "category": "electronics"
                    },
                    "quantity": 2
                  }
                ],
                "total": 159.98
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleOrderResponse"
                },
                "example": {
                  "order": {
                    "id": "ord-abc123",
                    "items": [
                      {
                        "product": {
                          "id": "prod-001",
                          "title": "Wireless Bluetooth Headphones",
                          "description": "Premium noise-cancelling wireless headphones with 30-hour battery life.",
                          "price": 79.99,
                          "image": "https://example.com/images/headphones.jpg",
                          "category": "electronics"
                        },
                        "quantity": 2
                      }
                    ],
                    "total": 159.98,
                    "date": "2026-02-23T14:30:00.000Z",
                    "status": "pending"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing items or total)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Items and total are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/reviews/{productId}": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get reviews for a product",
        "description": "Retrieves all reviews for a specific product. This endpoint does not require authentication.",
        "operationId": "getProductReviews",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the product to get reviews for",
            "schema": {
              "type": "string"
            },
            "example": "prod-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved reviews",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                },
                "example": {
                  "reviews": [
                    {
                      "id": "rev-xyz789",
                      "productId": "prod-001",
                      "userName": "Jane Doe",
                      "rating": 5,
                      "title": "Excellent sound quality!",
                      "body": "These headphones have amazing sound quality and the noise cancellation is top-notch. Battery life is as advertised.",
                      "date": "2026-02-20T10:15:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid product ID"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/payment/process": {
      "post": {
        "tags": ["Payment"],
        "summary": "Process a payment",
        "description": "Processes a simulated payment for an order. This endpoint does not connect to a real payment gateway; it simulates a successful payment transaction and returns a mock transaction ID.",
        "operationId": "processPayment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPaymentRequest"
              },
              "example": {
                "orderId": "ord-abc123",
                "amount": 159.98
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment successfully processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                },
                "example": {
                  "success": true,
                  "transactionId": "txn-9f8e7d6c5b4a",
                  "orderId": "ord-abc123",
                  "amount": 159.98,
                  "message": "Payment processed successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing orderId or amount, or invalid amount)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Order ID and amount are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the /api/auth/signup or /api/auth/signin endpoints. Include in the Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "description": "Represents a product in the catalog",
        "required": ["id", "title", "description", "price", "image", "category"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the product",
            "example": "prod-001"
          },
          "title": {
            "type": "string",
            "description": "Name of the product",
            "example": "Wireless Bluetooth Headphones"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the product",
            "example": "Premium noise-cancelling wireless headphones with 30-hour battery life."
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Price of the product in USD",
            "minimum": 0,
            "example": 79.99
          },
          "image": {
            "type": "string",
            "format": "uri",
            "description": "URL of the product image",
            "example": "https://example.com/images/headphones.jpg"
          },
          "category": {
            "type": "string",
            "description": "Category the product belongs to",
            "example": "electronics"
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "description": "Represents an item within an order",
        "required": ["product", "quantity"],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity of the product ordered",
            "minimum": 1,
            "example": 2
          }
        }
      },
      "Order": {
        "type": "object",
        "description": "Represents a customer order",
        "required": ["id", "items", "total", "date", "status"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the order",
            "example": "ord-abc123"
          },
          "items": {
            "type": "array",
            "description": "List of items in the order",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "total": {
            "type": "number",
            "format": "double",
            "description": "Total cost of the order in USD",
            "minimum": 0,
            "example": 159.98
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when the order was placed",
            "example": "2026-02-23T14:30:00.000Z"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "description": "The current status of an order",
        "enum": ["pending", "processing", "shipped", "delivered", "cancelled", "return_requested", "replacement_requested"],
        "example": "pending"
      },
      "Review": {
        "type": "object",
        "description": "Represents a product review submitted by a user",
        "required": ["id", "productId", "userName", "rating", "title", "body", "date"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the review",
            "example": "rev-xyz789"
          },
          "productId": {
            "type": "string",
            "description": "The ID of the product being reviewed",
            "example": "prod-001"
          },
          "userName": {
            "type": "string",
            "description": "Display name of the reviewer",
            "example": "Jane Doe"
          },
          "rating": {
            "type": "integer",
            "description": "Rating from 1 to 5 stars",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "title": {
            "type": "string",
            "description": "Brief title summarizing the review",
            "example": "Excellent sound quality!"
          },
          "body": {
            "type": "string",
            "description": "Full text of the review",
            "example": "These headphones have amazing sound quality and the noise cancellation is top-notch. Battery life is as advertised."
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when the review was submitted",
            "example": "2026-02-20T10:15:00.000Z"
          }
        }
      },
...
      "SeedProductsRequest": {
        "type": "object",
        "description": "Request body for seeding the product catalog",
        "required": ["action"],
        "properties": {
          "action": {
            "type": "string",
            "description": "The action to perform. Must be \"seed\".",
            "enum": ["seed"],
            "example": "seed"
          }
        }
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "description": "Request body for placing a new order",
        "required": ["items", "total"],
        "properties": {
          "items": {
            "type": "array",
            "description": "List of items to include in the order",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "total": {
            "type": "number",
            "format": "double",
            "description": "Total cost of the order in USD",
            "minimum": 0,
            "example": 159.98
          }
        }
      },
...
      "ProcessPaymentRequest": {
        "type": "object",
        "description": "Request body for processing a simulated payment",
        "required": ["orderId", "amount"],
        "properties": {
          "orderId": {
            "type": "string",
            "description": "The unique identifier of the order to pay for",
            "example": "ord-abc123"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "The payment amount in USD",
            "minimum": 0,
            "exclusiveMinimum": true,
            "example": 159.98
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message describing what went wrong",
            "example": "An unexpected error occurred"
          }
        }
      },
      "ProductListResponse": {
        "type": "object",
        "description": "Response containing a list of products",
        "required": ["products"],
        "properties": {
          "products": {
            "type": "array",
            "description": "Array of product objects",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "SingleProductResponse": {
        "type": "object",
        "description": "Response containing a single product",
        "required": ["product"],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "SeedProductsResponse": {
        "type": "object",
        "description": "Response after seeding the product catalog",
        "required": ["success", "message"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates the seed operation was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable message about the seeding result",
            "example": "Successfully seeded 20 products"
          }
        }
      },
...
      "OrderListResponse": {
        "type": "object",
        "description": "Response containing a list of orders",
        "required": ["orders"],
        "properties": {
          "orders": {
            "type": "array",
            "description": "Array of order objects",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "SingleOrderResponse": {
        "type": "object",
        "description": "Response containing a single order",
        "required": ["order"],
        "properties": {
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
...
      "ReviewListResponse": {
        "type": "object",
        "description": "Response containing a list of reviews for a product",
        "required": ["reviews"],
        "properties": {
          "reviews": {
            "type": "array",
            "description": "Array of review objects",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          }
        }
      },
...
      "PaymentResponse": {
        "type": "object",
        "description": "Response from a successful simulated payment",
        "required": ["success", "transactionId", "orderId", "amount", "message"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates the payment was processed successfully",
            "example": true
          },
          "transactionId": {
            "type": "string",
            "description": "Unique identifier for the payment transaction",
            "example": "txn-9f8e7d6c5b4a"
          },
          "orderId": {
            "type": "string",
            "description": "The order ID the payment was applied to",
            "example": "ord-abc123"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "The amount that was charged in USD",
            "example": 159.98
          },
          "message": {
            "type": "string",
            "description": "Human-readable message about the payment result",
            "example": "Payment processed successfully"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token is invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Authentication required"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}